db/ models/ repositories/ routes/ schemas/ utils/ moved to the main branch. Psycopg added into requirements.txt. Flask Cors installed. Backend server working ok
