+++ backend/requirements.txt
@@ -3,3 +3,4 @@ psycopg==3.1.18
 pytest==8.1.1
 pytest-playwright==0.4.4
 pytest-xprocess==0.23.0
+psycopg ==3.2.3
